Alvin branch

+++ src/headers/directory_operations.h
@@ -6,7 +6,9 @@
 #include "common.h"
 
 // Function declarations
-int create_directory(const char *dirname);
+// Note: create_directory returns uint16_t (inode number) or 0 on error
+// implementation is in dev.c
+uint16_t create_directory(const char *dirname);
 int delete_directory(const char *dirname);
 int change_directory(const char *dirname, SessionConfig *config);
 void list_directory_contents(const char *dirname, const SessionConfig *config);
